+++ CHANGELOG.rst
@@ -24,6 +24,8 @@ Changed
 
 - Re-enabled (collinear) spin polarization in transport
 
+- Option to take quasi-Newton steps in lBFGS (set as default)
+  
 Fixed
 -----